[FIX] mail: scroll-down to older messages
- Open a record with the chatter (a task for instance);
- Zoom your browser until the chatter is inline with the task (as it is
  the case on mobile)
- Write a big message.

Before this commit, the composer takes all the allowed place on the
screen, and we cannot see or scroll-down to the older messages.

Now, the behavior is the same as is done in mobile, the composer stays
sticky, but it will not take all the allowed place, and leave place to
see and scroll-down the older messages.

opw-3593032

closes odoo/odoo#143295

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>